Implement paging feature into List endpoints

+++ Models/ViewModels/PagingInfo.cs
@@ -8,6 +8,6 @@ namespace WarehouseManager.Models.ViewModels
         public int ItemsPerPage { get; set; }
         public int Page { get; set; }
         public int TotalPages =>
-            (int)Math.Ceiling((decimal)TotalItems / ItemsPerPage);
+            (int)Math.Ceiling((double)TotalItems / ItemsPerPage);
     }
 }
